Fix warning from ament auto static analysis
