+++ packml_msgs/package.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
   <name>packml_msgs</name>
   <version>0.0.0</version>
   <description>PackML messages</description>
   <maintainer email="dejanira_araiza@artc.a-star.edu.sg">Dejanira Araiza-Illan</maintainer>
-  <license>Apache License 2.0</license>
+  <license>Apache-2.0</license>
   <author >Shaun Edwards</author>
   <author>Dejanira Araiza-Illan (ROS 2 compatible)</author>
   
@@ -18,10 +18,10 @@
   <exec_depend>rosidl_default_runtime</exec_depend>
   <exec_depend>std_msgs</exec_depend>
 
-  <member_of_group>rosidl_interface_packages</member_of_group>
-
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
+
+  <member_of_group>rosidl_interface_packages</member_of_group>
 
   <export>
     <build_type>ament_cmake</build_type>
