+++ packml_ros/package.xml
@@ -14,15 +14,19 @@
   <build_depend>rclcpp</build_depend>
   <build_depend>packml_msgs</build_depend>
   <build_depend>packml_sm</build_depend>
-  <build_depend>Qt5</build_depend>
+  <build_depend>qtbase5-dev</build_depend>
 
   <exec_depend>rclcpp</exec_depend>
   <exec_depend>packml_msgs</exec_depend>
   <exec_depend>packml_sm</exec_depend>
-  <exec_depend>Qt5</exec_depend>
-
+  <exec_depend>libqt5-core</exec_depend>
+  <exec_depend>libqt5-gui</exec_depend>
+  <exec_depend>libqt5-opengl</exec_depend>
+  <exec_depend>libqt5-widgets</exec_depend>
+  
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
+  <test_depend>ament_cmake_gmock</test_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
