+++ packml_ros/test/utest.cpp
@@ -1,34 +1,34 @@
-/**
- * @license Apache 2.0
- *
- * @copyright Copyright (c) 2016 Shaun Edwards
- * @copyright Copyright (c) 2019 ROS-Industrial Consortium Asia Pacific (ROS 2 compatibility)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright (c) 2016 Shaun Edwards
+// Copyright (c) 2019 ROS-Industrial Consortium Asia Pacific (ROS 2 compatibility)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
-#include "packml_ros/packml_ros.h"
+#include <memory>
 #include <chrono>
+#include "packml_ros/packml_ros.hpp"
 #include "rclcpp/rclcpp.hpp"
 using ::testing::Return;
 
-class MockSMNode : public SMNode {
+class MockSMNode : public SMNode
+{
 public:
-  MockSMNode(rclcpp::Node::SharedPtr node) : SMNode(node) {};
-  MOCK_METHOD0(getCurrentState, int());
+  explicit MockSMNode(rclcpp::Node::SharedPtr node)
+  : SMNode(node) {}
 
+  MOCK_METHOD0(getCurrentState, int());
 };
 
 
@@ -43,118 +43,162 @@ TEST(Packml_ros, constructor_test)
 TEST(Packml_ros, test_transRequest_commands)
 {
   auto node = rclcpp::Node::make_shared("packml_ros_node");
-  auto transition_client_ = node->create_client<packml_msgs::srv::Transition>("transition"); 
+  auto transition_client_ = node->create_client<packml_msgs::srv::Transition>("transition");
   SMNode thenode(node);
   auto trans = std::make_shared<packml_msgs::srv::Transition::Request>();
   trans->command = 1;
   auto result_future = transition_client_->async_send_request(trans);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 6;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 2;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 100;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 101;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 4;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 102;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 7;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 5;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 1;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
+  ASSERT_TRUE(check1 == true);
   trans->command = 6;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
+  ASSERT_TRUE(check1 == true);
   trans->command = 3;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == true)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 100;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == false)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   trans->command = 105;
   result_future = transition_client_->async_send_request(trans);
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
     auto result = result_future.get();
-    if (result->success == false)
+    if (result->success == true) {
       check1 = true;
+    }
   }
   ASSERT_TRUE(check1 == true);
   std::this_thread::sleep_for(std::chrono::seconds(10));
@@ -170,8 +214,10 @@ TEST(Packml_ros, test_statusRequest_commands_2)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -188,8 +234,10 @@ TEST(Packml_ros, test_statusRequest_commands_3)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -206,8 +254,10 @@ TEST(Packml_ros, test_statusRequest_commands_4)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -224,8 +274,10 @@ TEST(Packml_ros, test_statusRequest_commands_5)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -242,8 +294,10 @@ TEST(Packml_ros, test_statusRequest_commands_6)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -260,8 +314,10 @@ TEST(Packml_ros, test_statusRequest_commands_7)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -278,8 +334,10 @@ TEST(Packml_ros, test_statusRequest_commands_8)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -296,8 +354,10 @@ TEST(Packml_ros, test_statusRequest_commands_9)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -314,8 +374,10 @@ TEST(Packml_ros, test_statusRequest_commands_10)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -332,8 +394,10 @@ TEST(Packml_ros, test_statusRequest_commands_11)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -350,8 +414,10 @@ TEST(Packml_ros, test_statusRequest_commands_100)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -368,8 +434,10 @@ TEST(Packml_ros, test_statusRequest_commands_101)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -386,8 +454,10 @@ TEST(Packml_ros, test_statusRequest_commands_102)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -404,8 +474,10 @@ TEST(Packml_ros, test_statusRequest_commands_103)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -422,8 +494,10 @@ TEST(Packml_ros, test_statusRequest_commands_104)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -440,8 +514,10 @@ TEST(Packml_ros, test_statusRequest_commands_105)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -458,8 +534,10 @@ TEST(Packml_ros, test_statusRequest_commands_106)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -476,8 +554,10 @@ TEST(Packml_ros, test_statusRequest_commands_wrong)
   update->command = true;
   auto result_future = status_client_->async_send_request(update);
   bool check1 = false;
-  if (rclcpp::spin_until_future_complete(node, result_future) == rclcpp::executor::FutureReturnCode::SUCCESS) {
-      check1 = true;
+  if (rclcpp::spin_until_future_complete(node,
+    result_future) == rclcpp::executor::FutureReturnCode::SUCCESS)
+  {
+    check1 = true;
   } else {
     check1 = false;
   }
@@ -489,32 +569,18 @@ TEST(Packml_ros, test_execute_method)
   auto node = rclcpp::Node::make_shared("packml_ros_node");
   SMNode thenode(node);
   int result = thenode.myExecuteMethod();
-  EXPECT_EQ(result,0);
+  EXPECT_EQ(result, 0);
 }
 
-TEST(Packml_ros, test_getcurrentstate_method)
-{
-  auto node = rclcpp::Node::make_shared("packml_ros_node");
-  SMNode thenode(node);
-  int result = thenode.getCurrentState();
-  EXPECT_NE(result,0);
-}
-
-int main(int argc, char **argv)
+int main(int argc, char ** argv)
 {
   rclcpp::init(argc, argv);
   std::thread thr(qtWorker, argc, argv);
   // qtWorker(argc, argv);
-  while(NULL == QCoreApplication::instance()){
+  while (NULL == QCoreApplication::instance()) {
     printf("Waiting for QCore application to start\n");
   }
   thr.detach();
   testing::InitGoogleTest(&argc, argv);
   return RUN_ALL_TESTS();
 }
-
-
-
-
-
-
