+++ packml_sm/include/packml_sm/transitions.hpp
@@ -1,31 +1,27 @@
-/**
- * @license Software License Agreement (Apache License)
- *
- * @copyright Copyright (c) 2016 Shaun Edwards
- * @copyright Copyright (c) 2019 ROS-Industrial Consortium Asia Pacific (ROS 2 compatibility)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright (c) 2016 Shaun Edwards
+// Copyright (c) 2019 ROS-Industrial Consortium Asia Pacific (ROS 2 compatibility)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 
-#ifndef PACKML_SM__TRANSITIONS_H_
-#define PACKML_SM__TRANSITIONS_H_
+#ifndef PACKML_SM__TRANSITIONS_HPP_
+#define PACKML_SM__TRANSITIONS_HPP_
 
 #include <QtGui>
 #include "QEvent"
 #include "QAbstractTransition"
-#include "packml_sm/common.h"
-#include "packml_sm/state.h"
+#include "packml_sm/common.hpp"
+#include "packml_sm/state.hpp"
 #include "rclcpp/rclcpp.hpp"
 
 
@@ -37,7 +33,6 @@ namespace packml_sm
 class CmdTransition : public QAbstractTransition
 {
 public:
-
   /**
   * @brief Function to transition to the clear state
   * @param from - original state
@@ -151,26 +146,24 @@ public:
   /**
   * @brief Function to define a transition from a triggering event
   * @param cmd_value - number of the new requested state
-  * @param name_value - name of the new requested state 
+  * @param name_value - name of the new requested state
   */
-  CmdTransition(const CmdEnum &cmd_value, const QString &name_value)
+  CmdTransition(const CmdEnum & cmd_value, const QString & name_value)
   : cmd(cmd_value), name(name_value) {}
 
 
   /**
   * @brief Function to define a transition from a triggering event
   * @param cmd_value - number of the new requested state
-  * @param name_value - name of the new requested state 
+  * @param name_value - name of the new requested state
   * @param from - original state
   * @param to - ending state
   */
   CmdTransition(
-    const CmdEnum &cmd_value, const QString &name_value,
+    const CmdEnum & cmd_value, const QString & name_value,
     PackmlState & from, PackmlState & to);
 
 protected:
-
-
   /**
   * @brief Function to check if a transition is valid
   * @param e - triggering event
@@ -182,7 +175,7 @@ protected:
   * @brief Function to trigger an action when the transition is happening
   * @param e - triggering event
   */
-  virtual void onTransition(QEvent * e) {std::cout<<e<<std::endl;}
+  virtual void onTransition(QEvent * e) {std::cout << e << std::endl;}
 
 
   /**
@@ -204,8 +197,6 @@ protected:
 class StateCompleteTransition : public QAbstractTransition
 {
 public:
-
-
   /**
   * @brief Constructor of the class
   */
@@ -226,7 +217,6 @@ public:
   virtual ~StateCompleteTransition() {}
 
 protected:
-
   /**
   * @brief Function to check if the transition is valid
   * @param e - triggering event
@@ -238,7 +228,7 @@ protected:
   * @brief Function to trigger an action when the transition is happening
   * @param e - triggering event
   */
-  virtual void onTransition(QEvent * e) {std::cout<<e<<std::endl;}
+  virtual void onTransition(QEvent * e) {std::cout << e << std::endl;}
 
 private:
 };
@@ -250,8 +240,6 @@ private:
 class ErrorTransition : public QAbstractTransition
 {
 public:
-
-
   /**
   * @brief Constructor of the class
   */
@@ -272,8 +260,6 @@ public:
   virtual ~ErrorTransition() {}
 
 protected:
-
-
   /**
   * @brief Function to check if the transition is valid
   * @param e - triggering event
@@ -285,8 +271,8 @@ protected:
   * @brief Function to trigger an action when the transition is happening
   * @param e - triggering event
   */
-  virtual void onTransition(QEvent * e) { std::cout<<e<<std::endl;}
+  virtual void onTransition(QEvent * e) {std::cout << e << std::endl;}
 };
 
 }  // namespace packml_sm
-#endif  // PACKML_SM__TRANSITIONS_H_
+#endif  // PACKML_SM__TRANSITIONS_HPP_
